+++ src/main/kotlin/com/softwareimprovementgroup/plugins/sigrid/toolWindow/SigridWindowFactory.kt
@@ -3,12 +3,14 @@ package com.softwareimprovementgroup.plugins.sigrid.toolWindow
 import com.intellij.icons.AllIcons
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.options.ShowSettingsUtil
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.intellij.ui.content.ContentFactory
 import com.softwareimprovementgroup.plugins.sigrid.SigridBundle
+import com.softwareimprovementgroup.plugins.sigrid.settings.SigridProjectSettingsConfigurable
 import com.softwareimprovementgroup.plugins.sigrid.toolWindow.panels.MaintainabilityPanel
 import com.softwareimprovementgroup.plugins.sigrid.toolWindow.panels.OpenSourceHealthPanel
 import com.softwareimprovementgroup.plugins.sigrid.toolWindow.panels.SecurityPanel
@@ -34,7 +36,12 @@ class SigridWindowFactory : ToolWindowFactory, DumbAware {
                 allPanels.forEach { it.loadData() }
             }
         }
-        toolWindow.setTitleActions(listOf(refreshAction))
+        val settingsAction = object : AnAction(SigridBundle["panel.settings.button"], null, AllIcons.General.Settings) {
+            override fun actionPerformed(e: AnActionEvent) {
+                ShowSettingsUtil.getInstance().showSettingsDialog(project, SigridProjectSettingsConfigurable::class.java)
+            }
+        }
+        toolWindow.setTitleActions(listOf(refreshAction, settingsAction))
 
         toolWindow.contentManager.addContent(contentFactory.createContent(maintainabilityPanel, SigridBundle["maintainability.tab"], false))
         toolWindow.contentManager.addContent(contentFactory.createContent(securityPanel, SigridBundle["security.tab"], false))